extract converting completion item logics
